+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Move `rivet version config-validate` to `rivet config validate`
 - Move `RIVET_CONCURRENT_UPLOADS` env var to CLI flag on appropriate commands (env var still works)
 - Streamline `rivet init` experience
+- Add `rivet token create public-namespace` command
 
 ### Fixed